Summary
This blog post taught us three things:

We learned that we can create a binary distribution by using the application plugin.
We learned how we can copy a file from the source directory to the target directory by using the Copy task.
We learned how we can add files to the binary distribution that is created by the application plugin.